move code in _backend to _matplotlib_backend
